requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
bcrypt==4.0.1
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
facenet-pytorch==2.5.3
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
bcrypt==4.0.1
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
facenet-pytorch==2.5.3
numpy<2
opencv-python==4.8.1.78
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.3.7
scikit-learn==1.3.0